neurotech_api/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.3
djangorestframework==3.15.2
fastapi==0.115.5
h11==0.14.0
idna==3.10
pg8000==1.31.2
psycopg2-binary==2.9.10
pycallibri_ecg==1.0.2
pydantic==2.10.1
pydantic_core==2.27.1
pyneurosdk2==1.0.14
python-dateutil==2.9.0.post0
requests==2.32.3
scramp==1.4.5
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
